Move FixIdentifier() from Utility to IdentifierFile
